fixed up the code for color sensor
