+++ src/main/java/com/team364/frc2020/subsystems/WheelOfFortune.java
@@ -11,9 +11,12 @@ import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import static com.team364.frc2020.States.*;
-import static com.team364.frc2020.commands.ColorSensor.*;
+
+import java.nio.ByteBuffer;
+
 import com.ctre.phoenix.motorcontrol.can.TalonSRX;
 import com.ctre.phoenix.motorcontrol.ControlMode;
+import edu.wpi.first.wpilibj.I2C;
 
 /**
  * The VM is configured to automatically run this class, and to call the functions corresponding to
@@ -26,11 +29,40 @@ public class WheelOfFortune extends SubsystemBase {
   private int max;
   private int rgb;
   private TalonSRX mWheelOfFortune;
+  private ByteBuffer buf = ByteBuffer.allocate(5);
+  /**
+   * 255, 0, 0 = red, 
+   * 0, 255, 255 = blue,  
+   * 0, 255, 0 = green, 
+   * 255, 255, 0 = yello 
+  */
+  public static int[] detectedColor;
+  I2C sensor;
   
   public WheelOfFortune() {
     mWheelOfFortune = new TalonSRX(20);
+    sensor = new I2C(I2C.Port.kOnboard, 0x39);
+    sensor.write(0x00,192);
   }
-  public void function(){
+  public void setDetectedColor(){
+    detectedColor = getColorArray();
+  }
+  public int[] getColorArray(){
+    return new int[] {red(), green(), blue()};
+  }
+  public int red(){
+    sensor.read(0x16, 3, buf);
+    return buf.get(0);
+  }
+  public int green(){ 
+    sensor.read(0x18, 2, buf);
+    return buf.get(0);
+  }
+  public int blue(){
+    sensor.read(0x1a, 2, buf);
+    return buf.get(0);
+  }
+  public void detectedColor(){
     if(getDetectedColor(230, 270, 0) && getDetectedColor(0, 25, 1) && getDetectedColor(0, 25, 3)) {
       colorState = ColorStates.RED;
     }else if(getDetectedColor(0, 25, 0) && getDetectedColor(230, 270, 1) && getDetectedColor(230, 270, 3)) {
@@ -42,21 +74,20 @@ public class WheelOfFortune extends SubsystemBase {
     }
     colorState = ColorStates.RED;
     switch (colorState) {
-        case RED:
-          mWheelOfFortune.set(ControlMode.Position, 20);
-        break;
-        case BLUE:
-          mWheelOfFortune.set(ControlMode.Position, 30);
-        break;
-        case GREEN:
-          mWheelOfFortune.set(ControlMode.Position, -10);
-        break;
-        case YELLO:
-          mWheelOfFortune.set(ControlMode.Position, -20);
-        break;
-        default: 
-          SmartDashboard.
-        break;
+      case RED:
+        mWheelOfFortune.set(ControlMode.Position, 20);
+      break;
+      case BLUE:
+        mWheelOfFortune.set(ControlMode.Position, 30);
+      break;
+      case GREEN:
+        mWheelOfFortune.set(ControlMode.Position, -10);
+      break;
+      case YELLO:
+        mWheelOfFortune.set(ControlMode.Position, -20);
+      break;
+      default: 
+      break;
 
     }   
   }
@@ -75,6 +106,9 @@ public class WheelOfFortune extends SubsystemBase {
       return true;
     } else {
       return false;
-    }
+    }   
+  }
+  public void moveWoF(double motorPower) {
+    mWheelOfFortune.set(ControlMode.PercentOutput, motorPower);
   }
 }
